Refactor hardware setup wizard into modular package
- Split monolithic 2555-line hardware_setup.py into keymasq/gui/wizards/hardware_setup/ package
- Modules: capture, dialog, discovery, identity, inventory, rows, templates, types
- Updated test imports accordingly
