+++ keymasq/gui/wizards/hardware_setup/__init__.py
@@ -1,0 +1,3 @@
+from keymasq.gui.wizards.hardware_setup.dialog import HardwareSetupDialog
+
+__all__ = ["HardwareSetupDialog"]
