+++ keymasq/gui/wizards/hardware_setup/types.py
@@ -1,0 +1,5 @@
+from typing import Any
+
+DetectedDevice = dict[str, Any]
+DetectedInterface = dict[str, Any]
+DetectedButton = dict[str, Any]
